🚚 move: 메뉴 store 정리 (#286)
