+++ src/components/menu-category/Categories.tsx
@@ -1,5 +1,5 @@
 import { useFetchCategories } from '@/hooks/menu/menu-category/useFetchCategories';
-import useCategoriesStore, { setCategories, setCategory } from '@/shared/store/menu-category';
+import useCategoriesStore, { setCategories, setCategory } from '@/shared/store/menu/menu-category';
 import useAuthState from '@/shared/store/session';
 import { useEffect } from 'react';
 import CategoryComponentPage from './Category';
@@ -8,7 +8,6 @@ const CategoriesComponentPage = () => {
   const storeId = useAuthState(state => state.storeId);
   const { data: categoreisData } = useFetchCategories(storeId ?? '');
   const category = useCategoriesStore(state => state.category);
-  const categories = useCategoriesStore(state => state.categories);
 
   useEffect(() => {
     if (categoreisData?.error === null) {
