+++ src/components/menu-category/modal/MenuCategoryModal.tsx
@@ -1,5 +1,5 @@
 import { useModal } from '@/hooks/service/ui/useModal';
-import useCategoriesStore from '@/shared/store/menu-category';
+import useCategoriesStore from '@/shared/store/menu/menu-category';
 import CategoryFormPage from '../Form';
 import styles from '../styles/modal.module.css';
 import CloseButton from '/public/icons/close.svg';
